maybe-positive is now correct
Previously, maybe-positive would never mark a matrix as non-positive since
it was testing the sign of eigenvalues' *magnitudes* (which are always
positive).

Since a symmetric matrix will have real-only eigenvalues, the correct test
is positivity of the real part of eigenvalues.

If this test fails, maybe-positive will appropriately mark the matrix's
:positive metadata.